Custom particles, alpha version (big commit #1)
What has been done :
 - The custom particle object exists and in instanciable (creatable) from the editor toolbox
 - The particles show up and are manageable with plenty of settings already (I intend to add even more if needed)
 - The particles spawn and death locations can be handled via a new particle area object (light green, has the same sparkle icon as the custom particle object itself)

TODO: Scripting (Every single setting should have getters, setters, faders and easers (when relevant) in Squirrel)
TODO: Make a particle editor alongside the level editor, because jeez, that's waaaaay too many settings
TODO: Store custom particle data in standalone file for easy editing reusability (and allow users to load custom particle data from files)

... and probably much more to do and fix :)

+++ src/object/rain_particle_system.cpp
@@ -190,7 +190,7 @@ void RainParticleSystem::update(float dt_sec)
     } else {
       m_angle_fade_time_remaining -= dt_sec;
       float progress = 1.f - m_angle_fade_time_remaining / m_angle_fade_time_total;
-      progress = static_cast<float>(m_angle_easing(progress));
+      progress = static_cast<float>(m_angle_easing(static_cast<double>(progress)));
       m_current_angle = progress * (m_target_angle - m_begin_angle) + m_begin_angle;
     }
     set_angle(m_current_angle);
